+++ packages/graphic-interface/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mlightcad/graphic-interface
 
+## 3.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @mlightcad/geometry-engine@3.1.3
+
 ## 3.1.3
 
 ### Patch Changes